Moved and added React components, upgraded API
React components were moved to the components folder, added new ones
(nav.js and getPsalms.js). Will try to get API working soon. Not sure if
it will look like this maybe I'll reuse the code in a bit different
manner but in general it should work the way I design it now.