Signup working, database user sample
Got basic signup working. It can be abused easily but that can be fixed
later. The code shoud be nice and clean and easy to understand what it
is doing. Included small update to testdb.db with two accounts
(admin:admin, user:user) with hashed password. Secret is not that
secret.

Next:
  - Login