Server routes cleanup, passport fix, react signup/login
Removed some obsolete server routes, made it a bit more clear. Tried to
connect react signup/login forms with passport. Only signup kinda works
but login doesn't respond on 'onSubmit'. Don't know what causes the
problem. The code should be the same as in the case of signup. Also when
I was checking passport I realized there is some ajor part missing. Need
to come up with a finish for it.

Next:
  - Fix the g*****n login form
  - Complete passport