Material-ui fix and more
This commit is crucial for this project. It contains most of the needed
changes that were required after moving from material-ui v0.20.0 to
v1.0.0-beta.30.

There is a huge difference between the code used by the components
that I needed to get used to but the code is much cleaner now.
I also did another changes to the codebase that won't be most likely
any noticeable but it makes me more comfortable with it.

Upcoming:
  - client-side authentication
  - backend authentication cleanup (maybe??)

Thinking about:
  - websocket implementation
  - token authentication (to prevent signout on reload)