Working websocket client verification + cleanup
Implemented the most important part of the process.
It depends on a ticket generated by the server based
on a token provided by the client.
It is possible to skip this and send the whole token.
I wasn't sure if that's a good thing to do because
client websocket makes get request.

I'll mot likely keep this pattern.