Better error handling, some name changed
I didn't like the previous approach (dialog box instead of the dialer),
so I used space in the appbar.
It provides 3 states: all is OK, an error happened, you got disconnected

There is problem with the disconnected/error logic. For that I'll
probably introduce a new value for better user status handling.

This will probably lead to implementing a interactive user profile.