[Client] Material-UI update, many updates to UI, auth state handling
Updated Material-UI from 1.0 to 1.5 and added dependency with material
design icons.
In DialForm I added some icons. AppBar got a nice menu with two buttons
that interact with the server (reconnect and relogin).

- Reconnect: Try to connect again to the websocket server.
- Relogin: Renew token.

HomePage now serves as a login page. In addition to that AuthPage is not
to be used until further development. In production there will be only
one account so SignUp is not needed.

I also added a new function to verify token.